% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chunks.R
\name{Chunks}
\alias{Chunks}
\title{Chunks}
\usage{
Chunks(vec, n.chunks, exclude = c("residuals", "predicted"))
}
\arguments{
\item{vec}{Any vector. Vector to split into chunks. No default.}

\item{n.chunks}{Integer vector of length 1. Number of chunks to split the vector into. No default.}

\item{exclude}{Character vector. Names of elements to remove from vector before splitting. Particularly relevant for use in package regone. Defaults to c("residuals", "predicted").}
}
\description{
Splitting vector into specified n.chunks.
}
