% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAddedVariablePlots.R
\name{CreateAddedVariablePlots}
\alias{CreateAddedVariablePlots}
\title{AddedVariablePlots}
\usage{
CreateAddedVariablePlots(nms, data, regressors, fit, save.plot = TRUE)
}
\arguments{
\item{nms}{Character vector. Regressors to be plotted against residuals. Each regressor is individually plotted against the externally studentized residuals. No default}

\item{data}{data.frame object. Data as prepared in "make.project.R". Regressor values, predicted response, and residuals. No default.}

\item{regressors}{Character vector. The regressor labels in data. No default.}

\item{fit}{lm object. Linear model fit to data. No default}

\item{save.plot}{Logical vector of length 1. If TRUE the added variable plots are saved to disk. Defaults to TRUE.}
}
\description{
Plot added variable plots. Much credit to: https://aosmith.rbind.io/2018/01/31/added-variable-plots/
}
