% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateBoxCoxPlot.R
\name{CreateBoxCoxPlot}
\alias{CreateBoxCoxPlot}
\title{CreateBoxCoxPlot}
\usage{
CreateBoxCoxPlot(fit, dir = "./", save.plot = TRUE, ...)
}
\arguments{
\item{fit}{lm object. Model fit to the data. No default.}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "."}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}

\item{...}{Additional arguments for the lindia::gg_boxcox() function.}
}
\description{
Simple wrapper for the lindia::gg_boxcox() function to maintain consistency in function layout. Adds option to save.plot.
}
