% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDfbetaPlot.R
\name{CreateDfbetaPlot}
\alias{CreateDfbetaPlot}
\title{CreateDfbetaPlot}
\usage{
CreateDfbetaPlot(nms, fit, critical.value = NULL, dir = ".",
  save.plot = TRUE)
}
\arguments{
\item{nms}{Character vector. Regresors from which to calculate dfbeta. No default.}

\item{fit}{lm object. Linear Model fit to the data. No default.}

\item{critical.value}{Numeric vector of length 1. Cutoff/critical value for DFBETA plots. Defaults to NULL, in which case no cut-off value is plotted.}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "."}

\item{save.plot}{Logical vector of length 1. If TRUE the DFBETA plots are saved to disk. Defaults to TRUE.}
}
\description{
Computes DFBETA for each parameter of the fitted model, plots the dfbetas as specified in nms, and optionally saves them to disk.
}
