% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFittedAgainstResidualsPlot.R
\name{CreateFittedAgainstResidualsPlot}
\alias{CreateFittedAgainstResidualsPlot}
\title{CreateFittedAgainstResidualspLot}
\usage{
CreateFittedAgainstResidualsPlot(data, dir = ".", save.plot = TRUE)
}
\arguments{
\item{data}{data.frame object. Data as prepared in "RunProject.R". Regressor values, predicted response, and residuals. No default.}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "."}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Creates and returns a plot with fitted against residuals plot.
}
