% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegressorAgainstResidualsPlot.R
\name{CreateRegressorAgainstResidualsPlot}
\alias{CreateRegressorAgainstResidualsPlot}
\title{CreateRegressorAgainstResidualsPlot}
\usage{
CreateRegressorAgainstResidualsPlot(data, nms, dir = "./", save.plot = TRUE)
}
\arguments{
\item{data}{data.frame object. Data as prepared in "make.project.R". Regressor values, predicted response, and residuals. No default.}

\item{nms}{Character vector. Regressors to be plotted against residuals. Each regressor is individually plotted against the externally studentized residuals. No default}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "./"}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Prepares data to plottable format, and plots as well as optionally saves faceted plot of regressors against externally studentized residuals.
}
