% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRegressorRegressorPlots.R
\name{CreateRegressorRegressorsPlots}
\alias{CreateRegressorRegressorsPlots}
\title{CreateRegressorRegressorsPlots}
\usage{
CreateRegressorRegressorsPlots(data, nms, dir = "./", save.plot = TRUE)
}
\arguments{
\item{data}{data.frame. Data as prepared in RunProject.R. No default.}

\item{nms}{Character vector. Variables from data to plot. No default.}

\item{dir}{Character vector of length 1. Directory in which to store the regressor against regressors plots. Defaults to "./"}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Simple wrapper for GGally::ggscatmat in order to save plots.
}
