% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTransformedQQPlot.R
\name{CreateTransformedQQPlot}
\alias{CreateTransformedQQPlot}
\title{CreateTransformedQQPlot}
\usage{
CreateTransformedQQPlot(
  fit,
  iv,
  dv = "density",
  data,
  lambda = 0.9,
  dir = "./",
  save.plot = TRUE
)
}
\arguments{
\item{fit}{lm object. Model fit to the data. No default.}

\item{iv}{Character vector. Independent variables. No default.}

\item{dv}{Character vector of length 1. Dependent variable. Defaults to "density"}

\item{data}{data.frame. Data that the model was fit to. No default.}

\item{lambda}{Numeric vector of length 1. Lambda to transform the dependent variable in fit. Default to 0.9.}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "."}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Uses the Box-Cox generated lambda to fit and display the QQ-plot for the newly fitted model. Note that
a manual lambda is input.
}
