% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateAnovaTable.R
\name{GenerateAnovaTable}
\alias{GenerateAnovaTable}
\title{GenerateAnovaTable}
\usage{
GenerateAnovaTable(fit, dir = ".", pretty.names = c("Sum sq",
  "Mean sq", "F value", "Pr(>F)"), ...)
}
\arguments{
\item{fit}{lm object. Linear model that has been fit to the data. No default.}

\item{pretty.names}{Character vector. Column names for the analysis of variance table. Defaults to c("Sum sq", "Mean sq", "F value", "Pr(>F)")}

\item{...}{Additional argument for anova().}
}
\description{
Simple wrapper that generates an Analysis of Variance (ANOVA) table using the built-in anova function, and saves that table to disk.
}
