% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAddedVariablePlots.R
\name{InvisibleAvPlots}
\alias{InvisibleAvPlots}
\title{InvisibleAvPlots}
\usage{
InvisibleAvPlots(fit, ...)
}
\arguments{
\item{fit}{lm object. LM model that has been fit to the data. No default.}

\item{...}{Additional arguments.}
}
\description{
Gets the data from the avPlots function to use with ggplot.
}
