% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InvisiblePlot.R
\name{InvisiblePlot}
\alias{InvisiblePlot}
\title{InvisiblePlot}
\usage{
InvisiblePlot(plot.call)
}
\arguments{
\item{plot.call}{Function call. The call from which the plot is generated. No default}
}
\description{
Gets the data from the plot call, but does not show the plot.
}
