% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAllPossibleRegression.R
\name{MakeTable}
\alias{MakeTable}
\title{MakeTable}
\usage{
MakeTable(tbl, labels.list, file.name, dir = ".", ...)
}
\arguments{
\item{tbl}{data.frame Data to be kable'd. No default.}

\item{labels.list}{List. Containing the columns to subset, and labelled with the pretty name of that column. No default}

\item{file.name}{Character vector of length 1. File name of the table. No default}

\item{...}{Additional arguments for kableExtra::kable.}
}
\description{
Runs kable over a table and writes to disk.
}
