% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMulticolinearityMeasures.R
\name{PlotVar}
\alias{PlotVar}
\title{PlotVar}
\usage{
PlotVar(plot.data, var = "vif")
}
\arguments{
\item{plot.data}{data.frame Use to plot the multicolinearity measures. No default.}

\item{var}{Character vector of length 1. The variable to plot. Defaults to "vif".}

\item{critical.value}{Numeric vector. Values that define whether there is possibly multicolinearity. Defautls to c(5, 10)}
}
\description{
Helper function to plot multicolinearity measures.
}
