% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunCrossValidation.R
\name{PredictRegsubsets}
\alias{PredictRegsubsets}
\title{PredictRegsubsets}
\usage{
PredictRegsubsets(dv, object, newdata, id, ...)
}
\arguments{
\item{object}{regsubsets object. Object returned from the leaps::regsubsets function. No default}

\item{newdata}{matrix. Data on which to make predictions. No default.}

\item{id}{Numeric vector of length 1. Model id. No default.}
}
\description{
Helper function for running variable selection and cross-validation. Source: https://uc-r.github.io/model_selection. See section 6.1 of Introduction to Statistical Learning.
}
