% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRESS.R
\name{PRESS}
\alias{PRESS}
\title{PRESS}
\usage{
PRESS(fit)
}
\arguments{
\item{fit}{lm object. LM model that has been fit to the data. No default.}
}
\description{
Function taken from @Tim's answer here: https://stats.stackexchange.com/questions/248603/how-can-one-compute-the-press-diagnostic
}
