% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAllPossibleRegression.R
\name{RunAllPossibleRegression}
\alias{RunAllPossibleRegression}
\title{RunAllPossibleRegression}
\usage{
RunAllPossibleRegression(fit, save.plot = TRUE)
}
\arguments{
\item{fit}{lm object. Full model fit to the data. No default.}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Runs "all possible regression" analysis using the olsrr::ols_step_best_subset() function, plots the
"relevant" measures and saves to disk.
}
