% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProject.R
\name{RunProject}
\alias{RunProject}
\title{RunProject}
\usage{
RunProject(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical vector of length 1. If TRUE messages are printed for each stage of the project, e.g. when running residual analysis. Defaults to TRUE.}
}
\description{
Run full code for project.
}
