% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateAddedVariablePlots.R
\name{Summarize}
\alias{Summarize}
\title{Summarize}
\usage{
Summarize(data, allvars, var)
}
\arguments{
\item{data}{data.frame. The data to summarize from. No default.}

\item{allvars}{Character vector. Names of regressors in data. No default.}

\item{var}{Character vector of length 1. Variable in data to maintain/regressor to plot in the added variable plot. No default.}
}
\description{
Creates set where all other variables but var are repeated medians. Var is maintained in original state. Much credit to: https://aosmith.rbind.io/2018/01/31/added-variable-plots/.
}
