% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCooksDistancePlot.R
\name{CreateCooksDistancePlot}
\alias{CreateCooksDistancePlot}
\title{CreateCooksDistancePlot}
\usage{
CreateCooksDistancePlot(fit, save.plot = TRUE)
}
\arguments{
\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}

\item{data}{data.frame object. Data as prepared in RunProject.R. Regressor values, predicted response, and residuals. No default.}
}
\description{
Simple wrapper to Compute Cook's distance and generate plot for it.
}
