% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateFittedAgainstActualPlot.R
\name{CreateFittedAgainstActualPlot}
\alias{CreateFittedAgainstActualPlot}
\title{CreateFittedAgainstActualPlot}
\usage{
CreateFittedAgainstActualPlot(nms, data, save.plot = TRUE)
}
\arguments{
\item{nms}{Character vector. Regressors to be plotted against residuals. Each regressor is individually plotted against the externally studentized residuals. No default}

\item{data}{data.frame object. Data as prepared in "make.project.R". Regressor values, predicted response, and residuals. No default.}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Plot  Much credit to: https://drsimonj.svbtle.com/visualising-residuals
}
