% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateQQPlot.R
\name{CreateQQPlot}
\alias{CreateQQPlot}
\title{CreateQQPlot}
\usage{
CreateQQPlot(data, dir = "./", save.plot = TRUE)
}
\arguments{
\item{data}{data.frame object. Data as prepared in RunProject.R. Regressor values, predicted response, and residuals. No default.}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "."}

\item{save.plot}{Logical vector of length 1. If TRUE the plot is saved to disk. Defaults to TRUE.}
}
\description{
Simple wrapper for making a qq-plot of the residuals with ggplot.
}
