% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateMulticolinearityMeasures.R
\name{GenerateMulticolinearityMeasures}
\alias{GenerateMulticolinearityMeasures}
\title{GenerateMulticolinearityMeasures}
\usage{
GenerateMulticolinearityMeasures(data, fit)
}
\arguments{
\item{data}{data.frame Data used to fit the model. Used for pair-wise correlation analysis and eigenvalue system analysis. No default.}

\item{fit}{lm object. Fitted model. No default}
}
\description{
Produces variance inflation factors, eigenvalue system analysis for the fitted model.
}
