% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunAllPossibleRegression.R
\name{RunAllPossibleRegression}
\alias{RunAllPossibleRegression}
\title{RunAllPossibleRegression}
\usage{
RunAllPossibleRegression(
  fit,
  performance.measures = c("adjr", "aic", "sbic", "cp"),
  pretty.labels = c("Adjusted R-squared", "AIC", "BIC", "C(p)"),
  dir = "./",
  save.plot = TRUE
)
}
\arguments{
\item{fit}{lm object. Full model fit to the data. No default.}

\item{performance.measures}{Character vector. Performance measures of candidate model to include in the performance table. Defaults to c("adjr", "aic", "sbic", "cp").}

\item{pretty.labels}{Character vector. Column names for performance measures in performance table. Defaults to c("Adjusted R-squared", "AIC", "BIC", "C(p)")}

\item{dir}{Character vector of lenght 1. Directory in which to store the plot. Ignored if save.plot is FALSE. Defaults to "./"}

\item{save.plot}{Logical vector of length 1. If TRUE then the plot is saved to disk. Defaults to TRUE}
}
\description{
Runs "all possible regression" analysis using the olsrr::ols_step_best_subset() function, plots the
"relevant" measures and saves to disk.
}
